Add d2l-ajax-framed
Uses frau-jwt to be able to work in iframes. This allowed removing a bunch of code, since frau-jwt takes care of XSRF tokens and caching itself.

This was added as a separate components for two reasons - first, as there doesn't seem to be a reliable way of determining if we are in an iframe, especially with WCs (the usual checks aren't reliable), and secondly because it greatly increases the size, due to the browserifying of frau-jwt and its dependencies.

In theory, we could probably use frau-jwt/framed for d2l-ajax-framed, and frau-jwt/local for the normal d2l-ajax, which wouldn't be overly large.