Performance improvements
While iron-ajax's _requestOptionsChanged observer observes a whole bunch of values, its docs specify that the request will only be retriggered if the URL, body, params change - otherwise you have to .generateRequest(). Because of this, I've simplified the observer in d2l-ajax, and also removed some default values (all of which match the underlying iron-ajax default value anyway).

Also, only add listener afterNextRender.

These changes reduced the rendering time (for 250 elements) in polyperf from ~210ms to ~180ms, or about a 15-20% improvement. Rendering an iron-ajax is about a 90-100ms ordeal, and an empty WC is about 40ms, meaning that realistically the fastest we could make d2l-ajax would be ~140ms.