Also add stopPropagation to error, if applicable
This was causing some headaches due to _getCachedAuthToken being called. In the "normal" error case (i.e. a request failing), we get an Event that we can stopPropagation on, but _getCachedAuthToken can throw a normal Error as an event (which can't stopPropagation) that also gets run through the onError handler. Simple solution is to just check if we can stop the event. If not, meh, since it's not going to propagate anyway.