Return Promises from generateRequest
iron-ajax supports a Promise-based API (although it's not really documented?), which makes doing multiple sequential requests much more fluent.

One slight weirdness is that for the framed request, we are forced to return the .completes, whereas in the non-framed case it works fine with just returning the result of generateRequest()... not sure why that is, but I've made both the `d2l-ajax` and `d2l-ajax-framed` use the same interface. It's slightly inconsistent with how `iron-ajax` itself is doing things, but it's also slightly nicer than having to remember to `request.completes.then` when using it.

Tested this with My Courses as well as Parent Portal UI and it works fine, so seems good for just a minor release.